pyproject.toml
[project]
name = "tg-agent-framework"
version = "0.1.0"
description = "Telegram + LangGraph Agent 快速开发框架"
requires-python = ">=3.11"

dependencies = [
    "langgraph>=0.3",
    "langchain-core>=0.3",
    "langchain-openai>=0.3",
    "aiogram>=3.4",
    "paramiko>=3.4",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[project.scripts]
tg-agent = "tg_agent_framework.cli.init:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
